[Fix] size_t to int

+++ ft_printf.h
@@ -21,12 +21,12 @@
 
 int		ft_printf(const char *s, ...);
 
-size_t	ft_putnbr_count(int n, size_t count);
-size_t	ft_putchar_count(char c, size_t count);
-size_t	ft_putstr_count(char *s, size_t count);
-size_t	ft_putui_count(unsigned int num, size_t count);
-size_t	ft_putnbr_base(unsigned long n, char *base, size_t count);
-size_t	ft_putaddress(unsigned long n, char *base, size_t count);
+int	ft_putnbr_count(int n, int count);
+int	ft_putchar_count(char c, int count);
+int	ft_putstr_count(char *s, int count);
+int	ft_putui_count(unsigned int num, int count);
+int	ft_putnbr_base(unsigned long n, char *base, int count);
+int	ft_putaddress(unsigned long n, char *base, int count);
 size_t	ft_strlen(const char *s);
 
 #endif
